Refactor SBML tests directory
