+++ .github/workflows/test_sbml_semantic_test_suite.yml
@@ -14,8 +14,8 @@ on:
       - python/sdist/amici/sbml_import.py
       - python/sdist/amici/import_utils.py
       - scripts/run-SBMLTestsuite.sh
-      - tests/sbml/testSBMLSuite.py
-      - tests/conftest.py
+      - tests/sbml/tests/testSBMLSuite.py
+      - tests/sbml/tests/conftest.py
   check_suite:
    types: [requested]
   workflow_dispatch:
