+++ scripts/run-SBMLTestsuite.sh
@@ -10,6 +10,7 @@ if [[ ! -d "tests/sbml-test-suite" ]]; then
 fi
 
 source venv/bin/activate
+export PYTHONPATH="${PYTHONPATH:+$PYTHONPATH:}."
 pip show pytest-xdist > /dev/null 2>&1 || pip install pytest-xdist
 pip install coverage pytest-cov
 
@@ -28,5 +29,5 @@ if [[ -d "${RESULT_DIR}" ]]; then
 fi
 mkdir "${RESULT_DIR}"
 
-pytest ./tests/sbml/testSBMLSuite.py $cases -rfsE -n auto \
+pytest ./tests/sbml/tests/testSBMLSuite.py $cases -rfsE -n auto \
   --cov=amici --cov-report=xml:"coverage_SBMLSuite.xml" --cov-append
